// superset-frontend/src/pages/Login/types.ts
/**
 * Licensed to the Apache Software Foundation (ASF) under one
 * or more contributor license agreements.  See the NOTICE file
 * distributed with this work for additional information
 * regarding copyright ownership.  The ASF licenses this file
 * to you under the Apache License, Version 2.0 (the
 * "License"); you may not use this file except in compliance
 * with the License.  You may obtain a copy of the License at
 *
 *   http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing,
 * software distributed under the License is distributed on an
 * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
 * KIND, either express or implied.  See the License for the
 * specific language governing permissions and limitations
 * under the License.
 */
import type { ReactNode } from 'react';

export interface LoginFormValues {
  username: string;
  password: string;
}

export interface RegisterFormValues {
  nickname: string;
  email: string;
  password: string;
  confirmPassword: string;
  verifyCode: string;
}

export enum AuthType {
  AuthOID = 0,
  AuthDB = 1,
  AuthLDAP = 2,
  AuthOauth = 4,
  AuthSAML = 5,
}

export type OAuthProvider = {
  name: string;
  icon: string;
};

export type OIDProvider = {
  name: string;
  url: string;
};

export type Provider = OAuthProvider | OIDProvider;

export type ActiveTab = 'login' | 'register';

export interface LoginPageProps {
  onLoginSuccess?: (values: LoginFormValues) => void;
  onRegisterSuccess?: (values: RegisterFormValues) => void;
  onSendVerifyCode?: (email: string) => Promise<void>;
  logoSlot?: ReactNode;
  hideBackground?: boolean;
}